Fixed warning

+++ source/game/game/game.cpp
@@ -4435,7 +4435,7 @@ namespace Game {
 				if (faction->getFactionDisconnectHandled() == false &&
 					(faction->getControlType() == ctNetwork)) {
 
-					if (aiInterfaces.size() <= i)
+					if (static_cast<int>(aiInterfaces.size()) <= i)
 						aiInterfaces.push_back(NULL);
 					if (aiInterfaces[i] == NULL && (server == NULL ||
 						server->isClientConnected(faction->getStartLocationIndex()) == false)) {
